requirements.txt
PyMuPDF
python-docx
requests
